# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConceptDistribution": {
        "description": "The professor's lens: how the class stands on one concept. Gaps fill\nin when the defense conversations of Phase 7 start naming them.",
        "properties": {
          "concept_id": {
            "title": "Concept Id",
            "type": "integer"
          },
          "developing": {
            "title": "Developing",
            "type": "integer"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "title": "Gaps",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "shaky": {
            "title": "Shaky",
            "type": "integer"
          },
          "solid": {
            "title": "Solid",
            "type": "integer"
          },
          "unseen": {
            "title": "Unseen",
            "type": "integer"
          }
        },
        "required": [
          "concept_id",
          "name",
          "unseen",
          "shaky",
          "developing",
          "solid"
        ],
        "title": "ConceptDistribution",
        "type": "object"
      },
...
      "ConceptMastery": {
        "description": "One concept in the student's picture. The trail ships with the state\n(spec section 9): the frontend never renders a bare label.",
        "properties": {
          "concept_id": {
            "title": "Concept Id",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_for_revisit": {
            "title": "Due For Revisit",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "m_eff": {
            "title": "M Eff",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "retention": {
            "title": "Retention",
            "type": "number"
          },
          "trail": {
            "items": {
              "$ref": "#/components/schemas/TrailLine"
            },
            "title": "Trail",
            "type": "array"
          }
        },
        "required": [
          "concept_id",
          "name",
          "description",
          "label",
          "m_eff",
          "retention",
          "due_for_revisit",
          "trail"
        ],
        "title": "ConceptMastery",
        "type": "object"
      },
...
      "CourseOut": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "created_at"
        ],
        "title": "CourseOut",
        "type": "object"
      },
      "CourseUpdate": {
        "properties": {
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "CourseUpdate",
        "type": "object"
      },
      "DistributionOut": {
        "properties": {
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/ConceptDistribution"
            },
            "title": "Concepts",
            "type": "array"
          }
        },
        "required": [
          "concepts"
        ],
        "title": "DistributionOut",
        "type": "object"
      },
...
      "GradeIn": {
        "description": "The professor's grade for a submission, already mapped to the spec's\n[0,1] scale.",
        "properties": {
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "score"
        ],
        "title": "GradeIn",
        "type": "object"
      },
      "GradeOut": {
        "properties": {
          "graded_at": {
            "title": "Graded At",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "submission_id": {
            "title": "Submission Id",
            "type": "integer"
          }
        },
        "required": [
          "submission_id",
          "score",
          "graded_at"
        ],
        "title": "GradeOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MasteryOut": {
        "properties": {
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/ConceptMastery"
            },
            "title": "Concepts",
            "type": "array"
          }
        },
        "required": [
          "concepts"
        ],
        "title": "MasteryOut",
        "type": "object"
      },
...
      "Problem": {
        "description": "The error body, also referenced from route response annotations so\nthe contract documents error shapes.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "RevisitConcept": {
        "description": "One revisit suggestion: the concept and one targeted variant, or none\nwhen no unattempted verified variant exists right now (the queue stays\ncalm either way).",
        "properties": {
          "concept_id": {
            "title": "Concept Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "variant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RevisitVariant"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "concept_id",
          "name",
          "variant"
        ],
        "title": "RevisitConcept",
        "type": "object"
      },
      "RevisitOut": {
        "properties": {
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/RevisitConcept"
            },
            "title": "Concepts",
            "type": "array"
          }
        },
        "required": [
          "concepts"
        ],
        "title": "RevisitOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Tirocinium API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}": {
      "delete": {
        "description": "Delete a course and drop its shard file. Refused while seats exist:\nseats carry student submission history, and deleting the course out from\nunder them would destroy it silently (decision 0013).",
        "operationId": "delete_course_api_v1_courses__course_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Course",
        "tags": [
          "courses"
        ]
      },
      "get": {
        "operationId": "get_course_api_v1_courses__course_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Course",
        "tags": [
          "courses"
        ]
      },
      "patch": {
        "operationId": "update_course_api_v1_courses__course_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Course",
        "tags": [
          "courses"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/mastery": {
      "get": {
        "operationId": "seat_mastery_api_v1_courses__course_id__mastery_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasteryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Seat Mastery",
        "tags": [
          "mastery"
        ]
      }
    },
    "/api/v1/courses/{course_id}/mastery/distribution": {
      "get": {
        "description": "The class's relationship to the material, per concept: label counts\nonly, no per-seat ranking (spec section 6). Seats that have produced no\nevidence on a concept count as unseen.",
        "operationId": "mastery_distribution_api_v1_courses__course_id__mastery_distribution_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mastery Distribution",
        "tags": [
          "mastery"
        ]
      }
    },
    "/api/v1/courses/{course_id}/revisit": {
      "get": {
        "description": "The concepts worth a fresh look, most faded first, each with one\ntargeted variant (spec section 5): the highest-weight published case\nstudy for the concept, excluding cases attempted in the last 48 hours,\ndrawing an unattempted servable variant from the pool.",
        "operationId": "revisit_queue_api_v1_courses__course_id__revisit_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisitOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revisit Queue",
        "tags": [
          "mastery"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/submissions/{submission_id}/grade": {
      "post": {
        "description": "The professor grades a submission: ground truth. One professor_grade\nevent per mapped concept (confidence 1.0), which triggers the\nsupersession replay inside the store, retracting the submission's\nautomatic events from the estimate (spec 4.6). Event insert, state\nreplay, and the grade column move in one writer transaction.",
        "operationId": "grade_submission_api_v1_courses__course_id__submissions__submission_id__grade_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "title": "Submission Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Grade Submission",
        "tags": [
          "mastery"
        ]
      }
    },
...
  }
}